refactor: extract writeChainInfo
